Create optimisation package
