+++ src/main/java/core/CoreParameters.java
@@ -1,6 +1,6 @@
 package core;
 
-import evaluation.TunableParameters;
+import evaluation.optimisation.TunableParameters;
 
 import java.util.Arrays;
 import java.util.Objects;
