+++ src/main/java/core/interfaces/IGameHeuristic.java
@@ -1,21 +1,6 @@
 package core.interfaces;
 
 import core.Game;
-import evaluation.TunableParameters;
-import org.json.simple.JSONArray;
-import org.json.simple.JSONObject;
-import org.json.simple.parser.JSONParser;
-import org.json.simple.parser.ParseException;
-import utilities.Utils;
-
-import java.io.FileNotFoundException;
-import java.io.FileReader;
-import java.io.IOException;
-import java.lang.reflect.Constructor;
-import java.lang.reflect.InvocationTargetException;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
 
 public interface IGameHeuristic {
