+++ src/main/java/evaluation/optimisation/GameEvaluator.java
@@ -1,4 +1,4 @@
-package evaluation;
+package evaluation.optimisation;
 
 import core.AbstractParameters;
 import core.AbstractPlayer;
