+++ src/main/java/games/coltexpress/ColtExpressParameters.java
@@ -2,7 +2,7 @@ package games.coltexpress;
 
 import core.AbstractParameters;
 import core.Game;
-import evaluation.TunableParameters;
+import evaluation.optimisation.TunableParameters;
 import games.GameType;
 import games.coltexpress.cards.ColtExpressCard;
 import utilities.Group;
