+++ src/main/java/games/explodingkittens/ExplodingKittensParameters.java
@@ -2,7 +2,7 @@ package games.explodingkittens;
 
 import core.AbstractParameters;
 import core.Game;
-import evaluation.TunableParameters;
+import evaluation.optimisation.TunableParameters;
 import games.GameType;
 import games.explodingkittens.cards.ExplodingKittensCard;
