+++ src/main/java/games/loveletter/LoveLetterHeuristic.java
@@ -4,7 +4,7 @@ import core.AbstractGameState;
 import core.CoreConstants;
 import core.components.PartialObservableDeck;
 import core.interfaces.IStateHeuristic;
-import evaluation.TunableParameters;
+import evaluation.optimisation.TunableParameters;
 import games.loveletter.cards.LoveLetterCard;
 import utilities.Utils;
