+++ src/main/java/games/loveletter/LoveLetterParameters.java
@@ -2,7 +2,7 @@ package games.loveletter;
 
 import core.AbstractParameters;
 import core.Game;
-import evaluation.TunableParameters;
+import evaluation.optimisation.TunableParameters;
 import games.GameType;
 import games.loveletter.cards.LoveLetterCard;
