+++ src/main/java/games/uno/UnoGameParameters.java
@@ -2,7 +2,7 @@ package games.uno;
 
 import core.AbstractParameters;
 import core.Game;
-import evaluation.TunableParameters;
+import evaluation.optimisation.TunableParameters;
 import games.GameType;
 
 import java.util.Arrays;
