+++ src/main/java/games/uno/UnoHeuristic.java
@@ -3,7 +3,7 @@ package games.uno;
 import core.AbstractGameState;
 import core.CoreConstants;
 import core.interfaces.IStateHeuristic;
-import evaluation.TunableParameters;
+import evaluation.optimisation.TunableParameters;
 
 public class UnoHeuristic extends TunableParameters implements IStateHeuristic {
