+++ src/main/java/games/virus/VirusGameParameters.java
@@ -2,7 +2,7 @@ package games.virus;
 
 import core.AbstractParameters;
 import core.Game;
-import evaluation.TunableParameters;
+import evaluation.optimisation.TunableParameters;
 import games.GameType;
 
 import java.util.Arrays;
