+++ src/main/java/players/learners/SVMEvaluator.java
@@ -5,8 +5,8 @@ import core.AbstractPlayer;
 import core.Game;
 import core.interfaces.IStateFeatureVector;
 import core.interfaces.ITunableParameters;
-import evaluation.ITPSearchSpace;
-import evaluation.ParameterSearch;
+import evaluation.optimisation.ITPSearchSpace;
+import evaluation.optimisation.ParameterSearch;
 import evodef.SearchSpace;
 import evodef.SolutionEvaluator;
 import games.GameType;
